Update gitignore and ignore 'html/dist/' directory
Files in this directory is generated from templates, it's therefore
little point having these in the repository.

There is still the remaining issue with files in 'wdoc/', these are
generated from files in 'html/dist/' and are those that actually gets
sent to the microcontroller.

The gitignore file was also simplified with regard to the '.vscode'
directory, this is now a single entry that excludes the entire
directory.