GH Actions: Create CI pipeline
We build a subset of the environments, at a later time, we should build
all of the possible environments and make them accessible for download.

To be certain that the html files is up to date, we still need to
generate the embedded files in 'wdoc/' when running GitHub Actions.