html: Move backup.html from extra/ to htmljs/src/
Since version 4.2 this has been included in the html files being
embedded in the binary file. It is therefore no longer appropriate to
have this file in the extra/ directory. It is moved to htmljs/src/ to
easier be integrated with 'grunt', and how we generate translated html
files in the repository.

Documentation has also been updated, a new section has been added to the
README in docs/.

The file is being handled as if it was being translated, but as of now
there isn't any translatable strings in the file.