html: Place generated embedded html files in wdoc/
When we generate the embedded html files we place them in the correct
directory immediately, we also won't clean the directory beforehand, only
overwrite old with new files.

This is an important step to be able to build artifacts with GitHub
Actions. It is now easier to generate these files during a CI run.

There is still a couple of files that is not being generated and this
prevents us from completely delete this directory.