Remove all references to set_manual_commit()
The implementation of this function has been disabled for several years.
The reasoning behind the functionality might have been valid once, spare
the flash for to many writes. I'm assuming it has been disabled because
of various other bugs, and it got to complicated to handle the write to
flash manually.