Change to unscoped enum for temperature Modes
This was originally several defines, and has now been changed to an
unscoped enum. A lot of the types has also been updated to 'Mode' for
increased type safety. This change leaked into the MQTT implementation
which I didn't like, though it was unavoidable.

For MQTT the value 'invalidMode' has been removed, is has been changed
to 'off'. The changes related to MQTT can be improved, the enum could
also be changed to a scoped enum.