[frontend] Have one external JS file shared for all HTML files
Combine all JavaScript code into a single file, except for 'dygraph'.
This will lower usage of flash and provide the possibility for the
browser to cache the JavaScript file, further increasing performance.

Were now using 'rollup' for creating ES6 modules.

'src/js/main.js' is the main entry for JS. This is also where all pages
calls on page load for initializing.

There is one caveat with this approach, we have translated strings in
the JavaScript. When the JavaScript was embedded in the HTML this was no
problem, it got translated at the same time as the HTML files. This is
no longer an option, so another approach must be found.