Fix loading of beer profiles in 'Control' page
There was several problems here. We initially sent an incomplete
filename to the frontend, missed the first character in the filename.
When it was received, it only showed the last character in the filename.
The result was that it was impossible to load the profile.

handleFileList() has been cleaned up for readability.