Validate JSON values before saving them
We're now validating the JSON received from the frontend, before saving
it. This makes certain that all information stored in the struct is
valid and it is therefore not needed to validate them before use.

This only applies to the struct SystemConfiguration which holds network
settings. Some cleanup, like better debug messages and returning early on
errors was also done at the same time.