Update types in System configuration form
The JSON object we send to the backend now contains the correct types,
previously all fields was a string, except the checkbox, this was a
number. Now the fields which contains number is sent as a number, while
the checkbox is being treated as a boolean value. This is also what the
backend sends back to the frontend during the GET request.

This also updates validation done in HTML, like max length and how high
or low allowed number. The backend should still have implemented a
better validation logic.