[frontend] Refactor to use one external CSS file
By changing to a single external CSS file we save space on the
embedded device since we no longer duplicate much of the CSS in an
embedded style section in each HTML file.

Change Sass plugin used when building the frontend, this requires Sass
as a module and not as a global accessible CLI.

We need to copy the .css file into each folder containing the translated
HTML files to have a functional developer setup as well as production
environment.

We only compress the .css file when we're testing on a device or for
production. This saves time when developing as well as prevents us from
handling content-encoding of the .css file.

The CSS styles requires some clean up before it's identical to what it
was before, this is specifically related to the control page.